refactor: skills sheet
